+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.dto.OrderTableRequest;
 import kitchenpos.table.domain.TableGroup;
 
 public class TableGroupRequest {
