+++ src/test/java/kitchenpos/order/OrderAcceptanceTest.java
@@ -6,7 +6,7 @@ import static kitchenpos.order.OrderFixture.주문;
 import static kitchenpos.order.OrderFixture.주문_목록_조회;
 import static kitchenpos.order.OrderFixture.주문_상태_수정;
 import static kitchenpos.product.ProductFixture.상품_등록;
-import static kitchenpos.table.TableFixture.주문_테이블_추가;
+import static kitchenpos.order.TableFixture.주문_테이블_추가;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 
@@ -23,8 +23,8 @@ import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderResponse;
-import kitchenpos.table.dto.OrderTableRequest;
-import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.order.dto.OrderTableRequest;
+import kitchenpos.order.dto.OrderTableResponse;
 import kitchenpos.product.dto.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
